Creating an Admin Product

+++ frontend/src/screens/ProductListScreen.js
@@ -4,9 +4,10 @@ import {Table, Button, Row, Col} from 'react-bootstrap'
 import {useDispatch, useSelector} from 'react-redux'
 import Message from '../components/Message'
 import Loader from '../components/Loader'
-import {listProducts, deleteProduct} from '../actions/productActions'
+import {listProducts, deleteProduct, createProduct} from '../actions/productActions'
+import {PRODUCT_CREATE_RESET} from '../constants/productConstants'
 
-const ProductListScreen = ({history, match}) => {
+const ProductListScreen = ({history}) => {
     const dispatch = useDispatch();
 
     const productList = useSelector((state) => state.productList);
@@ -15,21 +16,27 @@ const ProductListScreen = ({history, match}) => {
     const productDelete = useSelector((state) => state.productDelete);
     const {loading: loadingDelete, error: errorDelete, success: successDelete} = productDelete;
 
+    const productCreate = useSelector((state) => state.productCreate);
+    const {loading: loadingCreate, error: errorCreate, success: successCreate, product: createdProduct} = productCreate;
+
     const userLogin = useSelector((state) => state.userLogin);
     const {userInfo} = userLogin;
 
     useEffect(() => {
-        (userInfo && userInfo.isAdmin) ? dispatch(listProducts()) : history.push('/login')
-    }, [dispatch, history, userInfo, successDelete]);
+        dispatch({type: PRODUCT_CREATE_RESET});
+
+        if (!userInfo.isAdmin) history.push('/login');
+
+        (successCreate) ? history.push(`/admin/product/${createdProduct._id}/edit`) : dispatch(listProducts());
+
+    }, [dispatch, history, userInfo, successDelete, successCreate, createdProduct]);
 
     const deleteHandler = (id) => {
-        if (window.confirm('Are you sure')) {
-            dispatch(deleteProduct(id));
-        }
+        if (window.confirm('Are you sure')) dispatch(deleteProduct(id));
     };
 
-    const createProductHandler = (product) => {
-        console.log('create product');
+    const createProductHandler = () => {
+        dispatch(createProduct())
     };
 
     return <>
@@ -39,12 +46,14 @@ const ProductListScreen = ({history, match}) => {
             </Col>
             <Col className='text-right'>
                 <Button className='my-3' onClick={createProductHandler}>
-                    <i className='fas fa-plus'/>Create Product
+                    <i className='fas fa-plus'/> Create Product
                 </Button>
             </Col>
         </Row>
         {loadingDelete && <Loader/>}
         {errorDelete && <Message variant='danger'>{errorDelete}</Message>}
+        {loadingCreate && <Loader/>}
+        {errorCreate && <Message variant='danger'>{errorCreate}</Message>}
         {loading ? (
             <Loader/>
         ) : error ? (
